refactor: remove duplication on fetch speaker logic
