+++ src/2023/Speakers/Speakers2023.tsx
@@ -20,7 +20,7 @@ import {
 import webData from "../../data/2023.json";
 import Button from "../../components/UI/Button";
 import {gaEventTracker} from "../../components/analytics/Analytics";
-import {useFetchSpeakers} from "./UseFetchSpeakers";
+import {useFetchSpeakers} from "../../hooks/useFetchSpeakers";
 import * as Sentry from "@sentry/react";
 import {ISpeaker} from "../../types/speakers";
 
@@ -41,7 +41,7 @@ const Speakers2023: FC<React.PropsWithChildren<unknown>> = () => {
   const isBetween = (startDay: Date, endDay: Date): boolean =>
     startDay < new Date() && endDay > today;
 
-  const { error, data, isLoading } = useFetchSpeakers();
+  const { error, data, isLoading } = useFetchSpeakers("2023");
 
   if (error) {
     Sentry.captureException(error);
