+++ src/2023/TalkDetail/TalkDetailContainer2023.tsx
@@ -7,7 +7,7 @@ import {useParams} from "react-router";
 import conferenceData from "../../data/2023.json";
 import {useFetchTalksById} from "../Talks/UseFetchTalks";
 import * as Sentry from "@sentry/react";
-import {useFetchSpeakers} from "../Speakers/UseFetchSpeakers";
+import {useFetchSpeakers} from "../../hooks/useFetchSpeakers";
 import {Session} from "../../types/sessions";
 import TalkDetail from "./TalkDetail";
 import {ISpeaker} from "../../types/speakers";
@@ -19,7 +19,7 @@ const StyledContainer = styled.div`
 const TalkDetailContainer2023: FC<React.PropsWithChildren<unknown>> = () => {
   const { id } = useParams<{ id: string }>();
   const { isLoading, error, data } = useFetchTalksById(id!);
-  const { data: speakerData } = useFetchSpeakers();
+  const { data: speakerData } = useFetchSpeakers("2023");
 
   const getTalkSpeakers = (
     data: Session[] | undefined,
